refactor(arena): split the 962-line arena.ts into five modules
Spec §13 caps a file at 400 lines; arena.ts was 962 and had been the
plan's known debt since it was 780. It is now 181 lines of orchestration
over four focused modules.

- arenaWorld: ArenaWorld/ArenaSetup and the shared constants. Imports no
  sibling, so it sits at the bottom without risking a cycle.
- boarding: Instappen/Uitstappen and the edge-triggered buttons. Combat
  and roster both throw a player out of a car, so exitVehicle lives here
  rather than in either of them — which is what made the split a cycle
  risk and kept it out of the behaviour PR.
- movement: cars, walking, riding and the collision passes.
- combat: firing, bullets, hits and explosions.
- roster: create, join, leave, respawn, eject-on-death.

Pure code movement. Normalising imports and the added export keywords,
the before and after are line-for-line identical (828 either side), so
the only behaviour question is module wiring, which tsc settles.
